Move create_samples() function into preprocessing.py
